+++ src/main/java/com/brand/blockus/data/providers/BlockusModelProvider.java
@@ -117,9 +117,9 @@ public class BlockusModelProvider extends FabricModelProvider {
         this.registerPillar(modelGenerator, BlockusBlocks.DRIPSTONE_PILLAR);
 
         // Tuff
-        modelGenerator.registerSimpleCubeAll(BlockusBlocks.CHISELED_TUFF);
+        modelGenerator.registerSimpleCubeAll(BlockusBlocks.CARVED_TUFF);
         modelGenerator.registerSimpleCubeAll(BlockusBlocks.CRACKED_TUFF_BRICKS);
-        this.registerButtonAndPressurePlate(modelGenerator, BlockusBlocks.POLISHED_TUFF_PRESSURE_PLATE, BlockusBlocks.POLISHED_TUFF_BUTTON, BlockusBlocks.POLISHED_TUFF.block);
+        this.registerButtonAndPressurePlate(modelGenerator, BlockusBlocks.POLISHED_TUFF_PRESSURE_PLATE, BlockusBlocks.POLISHED_TUFF_BUTTON, Blocks.POLISHED_TUFF);
         modelGenerator.registerSouthDefaultHorizontalFacing(TexturedModel.TEMPLATE_GLAZED_TERRACOTTA, BlockusBlocks.TUFF_CIRCULAR_PAVING);
         this.registerPillar(modelGenerator, BlockusBlocks.TUFF_PILLAR);
         modelGenerator.registerSimpleCubeAll(BlockusBlocks.HERRINGBONE_TUFF_BRICKS);