+++ changelog.md
@@ -1,3 +1,5 @@
-### Updated to **1.20.5**
-* Replace #patterned_wool by #all_patterned_wools in #wool block tag
+### Updated to **1.21**
+* Removed modded Polished Tuff & Tuff Bricks blocks.
+* Renamed modded Chiseled Tuff into Carved Tuff.
+* Modded Tuff variants recipes now use vanilla Tuff blocks.